Just a quick fix to restore functionality to example(getGene)
Inside checkWrapperArgs(), martFilters(mart) is now returning
something other than what it used to.  Now its a dataframe which is
not easily coerced to a list.  listFilters() still works the same
though and can also work here in our pinch.  So that is what it does
for now to restore functionality.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/biomaRt@38839 bc3139a8-67e5-0310-9ffc-ced21a209358